requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.0.6
django-extensions==3.2.3
django-ninja==1.2.2
pydantic==2.8.22
pydantic_core == 2.8.22
sqlparse==0.5.0
typing_extensions==4.12.0
tzdata==2024.1

